Fix to Phantasmal_Killer so it is removed on concentration drop by making Remove_Paired_Effect recognize UUID of active effect (likely also fixes many other macros). Also adds a Convenient Effect (CE) description and chat bubble.

+++ Utility_Macros/Remove_Paired_Effect/Remove_Paired_Effect.js
@@ -1,4 +1,4 @@
-const MACRONAME = "Remove_Paired_Effect.0.4.js"
+const MACRONAME = "Remove_Paired_Effect.0.5.js"
 /*****************************************************************************************
  * Remove the specified  effect. 
  * 
@@ -16,6 +16,7 @@ const MACRONAME = "Remove_Paired_Effect.0.4.js"
  * 06/04/22 0.2 Removal of uneccesary comments and such
  * 06/09/22 0.3 Accept either a token or actor id
  * 07/07/22 0.4 Added (preferred) option of supplying UUID of the effect
+ * 07/28/22 0.5 Detect UUID as args[2] and just use that
  *****************************************************************************************/
 jez.log(`============== Starting === ${MACRONAME} =================`);
 for (let i = 0; i < args.length; i++) jez.log(`  args[${i}]`, args[i]);
@@ -26,33 +27,40 @@ if (args[0] === "off") {  // Only execute this when removed from Token
     // https://gitlab.com/tposney/dae/-/issues/319
     // await jez.wait(Math.floor(Math.random() * 500)) // Fixed in DAE 0.10.22
     //------------------------------------------------------------------------------------
-    if (args.length === 3) {
+    if (args.length === 3) {                                        // 1st Arg is UUID
         const EFFECT = await fromUuid(args[1])
-        jez.log("EFFECT", EFFECT)
+        jez.log("EFFECT from arg 1", EFFECT)
         if (EFFECT) EFFECT.delete()
+        return
     }
-    else {
-        if (args.length === 4) {
-            let fetchedToken = canvas.tokens.placeables.find(ef => ef.id === args[1])
-            jez.log(`fetchedToken ${fetchedToken?.name}`, fetchedToken)
-            if (fetchedToken) {
-                let existingEffect = await fetchedToken.actor.effects.find(i => i.id === args[2]);
-                if (existingEffect) await existingEffect.delete()
-            }
-            else {
-                let fetchedActor = canvas.tokens.placeables.find(ef => ef.data.actorId === args[1]).actor
-                jez.log(`fetchedActor ${fetchedActor?.name}`, fetchedActor)
-                if (fetchedActor) {
-                    jez.log("fetchedActor.effects", fetchedActor.effects)
-                    let existingEffect = await fetchedActor.effects.find(i => i.id === args[2]);
-                    if (existingEffect) {
-                        jez.log("existingEffect to be deleted", existingEffect)
-                        await existingEffect.delete()
-                    }
+    if (args.length === 4 && args[2].includes(".ActiveEffect.")) {  // 2nd Arg is UUID
+        const EFFECT = await fromUuid(args[2])
+        jez.log("EFFECT from arg 2", EFFECT)
+        if (EFFECT) EFFECT.delete()
+        return
+    }
+    if (args.length === 4 && args[2].length === 16) {                // 2nd Arg is Id
+        let fetchedToken = canvas.tokens.placeables.find(ef => ef.id === args[1])
+        jez.log(`fetchedToken ${fetchedToken?.name}`, fetchedToken)
+        if (fetchedToken) {
+            let existingEffect = await fetchedToken.actor.effects.find(i => i.id === args[2]);
+            if (existingEffect) await existingEffect.delete()
+        }
+        else {
+            let fetchedActor = canvas.tokens.placeables.find(ef => ef.data.actorId === args[1]).actor
+            jez.log(`fetchedActor ${fetchedActor?.name}`, fetchedActor)
+            if (fetchedActor) {
+                jez.log("fetchedActor.effects", fetchedActor.effects)
+                jez.log("seeking id", args[2])
+                let existingEffect = await fetchedActor.effects.find(i => i.id === args[2]);
+                if (existingEffect) {
+                    jez.log("existingEffect to be deleted", existingEffect)
+                    await existingEffect.delete()
                 }
+                else jez.log("effect not found for", args[2])
             }
         }
-        else jez.badNews(`Bad Arguments received by Remove_Paired_Effect, count ${args.length}`,"error")
+        return
     }
+    jez.badNews(`Bad Arguments received by Remove_Paired_Effect, count ${args.length}`, "error")
 }
-jez.log(`============== Finished === ${MACRONAME} =================`);
